Add clear and delete functionality
The user can now clear the display and corresponding variables using the
clear button, as well as delete the last character typed by using the
delete button.

Also, results that are floats with more that the specified number of
decimal places (DECIMALPLACES) will be rounded to that number of digits.